+++ wca.share/wca.share.application/Features/Funcionarios/Commands/FuncionarioUpdateCommandHandle.cs
@@ -21,7 +21,7 @@ namespace wca.share.application.Features.Funcionarios.Commands
         DateTime? DataDemissao = null,
         string? Email = null,
         int? DDDCelular = null,
-        int? NumeroCelular = null,
+        double? NumeroCelular = null,
         string? eSocialMatricula = null
     ) : IRequest<ErrorOr<FuncionarioResponse>>;
     internal class FuncionarioUpdateCommandHandle : IRequestHandler<FuncionarioUpdateCommand, ErrorOr<FuncionarioResponse>>
